% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{batch_correction}
\alias{batch_correction}
\title{batch_correction}
\usage{
batch_correction(dataset, pheno, covariate = NULL, batch)
}
\arguments{
\item{dataset}{input dataframe storing standardized data for batch correction - make sure the rows are appropriately ordered}

\item{pheno}{input dataframe storing phenotype data for batch correction - make sure the rows are appropriately ordered}

\item{covariate}{input covariate names in pheno whose variations need to be retained}

\item{batch}{input batch column name in pheno}
}
\value{
a dataframe with batch corrected data after ComBat
}
\description{
Batch correction using ComBat
}
