% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{batch_evaluation}
\alias{batch_evaluation}
\title{batch_evaluation}
\usage{
batch_evaluation(dataset, pheno, batch_col)
}
\arguments{
\item{dataset}{input dataframe storing data for batch effect evaluation - make sure the rows are appropriately ordered}

\item{pheno}{input dataframe storing phenotype data - make sure the rows are appropriately ordered}

\item{batch_col}{column name in pheno dataframe that stores batch information}
}
\value{
a list of different evaluation metrics for batch effect: if input data is on original scale, use cv; if standardized, use sd
}
\description{
Evaluate batch effects across different batches
}
