% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biphasic_fit_bayesian.R
\name{biphasic_fit_bayesian}
\alias{biphasic_fit_bayesian}
\title{biphasic_fit_bayesian}
\usage{
biphasic_fit_bayesian(data)
}
\arguments{
\item{data}{dataset prepared by prepare_data}
}
\value{
a large list of result lists ordered by sample name
}
\description{
biphasic_fit_bayesian
}
\examples{
single_data <- data.frame(time = c(0, 15, 30, 60, 120), F_t = c(40,26,19,15,12))
result_list <- biphasic_fit_bayesian(single_data)

}
