% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biphasic_fit_nls.R
\name{biphasic_fit_nls}
\alias{biphasic_fit_nls}
\title{biphasic_fit_nls}
\usage{
biphasic_fit_nls(data, initial_params)
}
\arguments{
\item{data}{prepared data}

\item{initial_params}{initial values for fitting}
}
\value{
a list of model fit par, estimates and plot
}
\description{
biphasic_fit_nls
}
\examples{
single_data <- data.frame(time = c(0, 15, 30, 60, 120), F_t = c(40,26,19,15,12))
params <- c(40,0,0.1,0.01)
result_list = biphasic_fit_nls(single_data, params)

}
