% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_half_life.R
\name{calculate_half_life}
\alias{calculate_half_life}
\title{calculate_half_life}
\usage{
calculate_half_life(F, S, k_f, k_s)
}
\arguments{
\item{F}{fast phase span}

\item{S}{slow phase span}

\item{k_f}{fast phase exp decay par}

\item{k_s}{slow phase exp decay par}
}
\value{
a list with overall, fast phase, slow phase half-life estimates
}
\description{
calculate_half_life
}
\examples{
half_life <- calculate_half_life(40,0,0.1,0.01)
}
