% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comet_looping.R
\name{comet_list_to_df}
\alias{comet_list_to_df}
\title{comet_list_to_df}
\usage{
comet_list_to_df(results_list, method = "adaptive")
}
\arguments{
\item{results_list}{A list generated from \code{loop_comet_data()}.}

\item{method}{Character; method used to fit: "nls", "bayes", or "adaptive".}
}
\value{
A data frame summarizing fitted parameters per sample.
}
\description{
Summarize list of comet fitting results into a single data frame.
}
