% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_results_to_dataframe.R
\name{convert_results_to_dataframe}
\alias{convert_results_to_dataframe}
\title{convert_results_to_dataframe}
\usage{
convert_results_to_dataframe(results_list)
}
\arguments{
\item{results_list}{result list generated by loop_data}
}
\value{
summary dataframe one row per sample
}
\description{
convert_results_to_dataframe
}
\examples{
in_data <- data.frame(Sample = c("001","002","003"),
c_0 = c(40,50,60), c_15=c(25,27,30), c_30=c(20,21,22), c_60=c(15,17,18), c_120=c(12,11,10))
sample_list <- loop_data(in_data)
sample_dat <- convert_results_to_dataframe(sample_list)
}
