% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_biphasic_bayes.R
\name{fit_biphasic_bayes}
\alias{fit_biphasic_bayes}
\title{fit_biphasic_bayes}
\usage{
fit_biphasic_bayes(data)
}
\arguments{
\item{data}{dataset prepared by prepare_data, a df with with 'time' and 'F_t' columns}
}
\value{
a large list of result lists ordered by sample name
}
\description{
fit_biphasic_bayes
}
