% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_biphasic_nls.R
\name{fit_biphasic_nls}
\alias{fit_biphasic_nls}
\title{fit_biphasic_nls}
\usage{
fit_biphasic_nls(data, initial_params)
}
\arguments{
\item{data}{prepared comet data}

\item{initial_params}{vector of initial parameters for the model: F, S, k_f, k_s}
}
\value{
a list of model fit par, estimates and plot
}
\description{
Fit biphasic model using non-linear least squares (nlsLM)
}
