% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_and_stats.R
\name{fit_model_and_stats}
\alias{fit_model_and_stats}
\title{fit_model_and_stats}
\usage{
fit_model_and_stats(data, model_function, initial_params)
}
\arguments{
\item{data}{prepared data}

\item{model_function}{model formula}

\item{initial_params}{initial values for fitting}
}
\value{
a list of model fit par, estimates and plot
}
\description{
fit_model_and_stats
}
\examples{
out_data <- data.frame(time = c(0, 15, 30, 60, 120), F_t = c(40,26,19,15,12))
params <- c(40,0,0.1,0.01)
result_list = fit_model_and_stats(out_data, model, params)

}
