% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_monophasic_model.R
\name{fit_monophasic_model}
\alias{fit_monophasic_model}
\title{fit_monophasic_model}
\usage{
fit_monophasic_model(data, initial_params)
}
\arguments{
\item{data}{prepared data for monophasic model}

\item{initial_params}{a vector of two values c(F, k)}
}
\value{
list of results
}
\description{
fit_monophasic_model
}
\examples{
out_data <- data.frame(time = c(0, 15, 30, 60, 120), F_t = c(48,40,19,15,12))
mono_list <- fit_monophasic_model(out_data, list(F = 48, k = 0.04))
}
