% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_2_df.R
\name{list_2_df}
\alias{list_2_df}
\title{list_2_df}
\usage{
list_2_df(results_list, method)
}
\arguments{
\item{results_list}{result list generated by loop_data}

\item{method}{method used to fit the model}
}
\value{
summary dataframe one row per sample
}
\description{
list_2_df
}
\examples{
in_data <- data.frame(Sample = c("001","002","003"),
c_0 = c(40,50,60), c_15=c(25,27,30), c_30=c(20,21,22), c_60=c(15,17,18), c_120=c(12,11,10))
sample_list <- loop_data(in_data, "nls")
sample_dat <- list_2_df(sample_list, "nls")

}
