% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comet_looping.R
\name{loop_comet_data}
\alias{loop_comet_data}
\title{loop_comet_data}
\usage{
loop_comet_data(dataset_name, method = "adaptive")
}
\arguments{
\item{dataset_name}{full dataset name}

\item{method}{Character; choose between "nls", "bayes", or "adaptive" (default = "adaptive").}
}
\value{
a large list of result lists ordered by sample name
}
\description{
Loop through each sample in the dataset and fit the biphasic model.
}
