% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_data.R
\name{loop_data}
\alias{loop_data}
\title{loop_data}
\usage{
loop_data(dataset_name)
}
\arguments{
\item{dataset_name}{full dataset name}
}
\value{
a large list of result lists ordered by sample name
}
\description{
loop_data
}
\examples{
in_data <- data.frame(Sample = c("001","002","003"),
c_0 = c(40,50,60), c_15=c(25,27,30), c_30=c(20,21,22), c_60=c(15,17,18), c_120=c(12,11,10))
sample_list <- loop_data(in_data)
}
