% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{model}
\usage{
model(t, F, S, k_f, k_s)
}
\arguments{
\item{t}{time}

\item{F}{fast phase span}

\item{S}{slow phase span}

\item{k_f}{fast phase par}

\item{k_s}{slow phase par}
}
\value{
to be used internally
}
\description{
model
}
