% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameters.R
\name{optimize_parameters}
\alias{optimize_parameters}
\title{optimize_parameters}
\usage{
optimize_parameters(data, initial_params)
}
\arguments{
\item{data}{prepared data with required time and F_t columns}

\item{initial_params}{a vector of numeric values; for biphasic exponential decay model input c(F, S, k_f, k_s)}
}
\value{
optimized parameters in c(F, S, k_f, k_s)
}
\description{
Optimize parameters for the biphasic exponential decay model using non-linear least squares.
}
