% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{RE_to_zscore}
\alias{RE_to_zscore}
\title{RE_to_zscore}
\usage{
RE_to_zscore(dataset, var_inverse = NULL)
}
\arguments{
\item{dataset}{input dataframe storing raw reporter expression data - make sure the data is appropriately ordered}

\item{var_inverse}{input pathway names with inverse relationship between raw reporter expression and pathway activity}
}
\value{
a dataframe with standardized data after natural log transformation and z-score normalization
}
\description{
RE_to_zscore
}
\examples{
dat_og <- data.frame(UG = c(0.05,0.08,0.10), NHEJ = c(23,16,19), HR = c(2,4,6))
dataset_std <- RE_to_zscore(dat_og, c("UG"))
}
