% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_parrallel.R
\name{setup_parallel}
\alias{setup_parallel}
\title{setup_parallel}
\usage{
setup_parallel(workers = future::availableCores())
}
\arguments{
\item{workers}{Number of CPU cores to use. Default = all available cores.}
}
\value{
None (sets options globally).
}
\description{
Setup parallelization for faster Bayesian model fitting and adaptive spike modeling.
}
