% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_evaluation.R
\name{batch_evaluation}
\alias{batch_evaluation}
\title{batch_evaluation}
\usage{
batch_evaluation(dataset, pheno, batch_col)
}
\arguments{
\item{dataset}{input dataframe storing data for batch effect evaluation - make sure the rows are appropriately ordered}

\item{pheno}{input dataframe storing phenotype data - make sure the rows are appropriately ordered}

\item{batch_col}{column name in pheno dataframe that stores batch information}
}
\value{
a list of different evaluation metrics for batch effect: if input data is on original scale, use cv; if standardized, use sd
}
\description{
batch_evaluation
}
\examples{
dat_og <- data.frame(UG = c(0.05,0.08,0.10), NHEJ = c(23,16,19), HR = c(2,4,6))
list_std <- RE_to_zscore(dat_og, c("UG"))
dat_std <- list_std$dataset_zscore
pheno <- data.frame(Sample = c("001","002","003"),
age = c(34,36,57), sex = c("M","F","M"), batch = c("A","B","A"))
list = batch_evaluation(dataset, pheno, "batch")

}
