% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comet_looping.R
\name{comet_list_to_df}
\alias{comet_list_to_df}
\title{comet_list_to_df}
\usage{
comet_list_to_df(results_list, method = "adaptive")
}
\arguments{
\item{results_list}{Output from \code{loop_comet_data()}}

\item{method}{"nls", "bayes", or "adaptive"}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        Tibble: one row per sample, one block of
                <estimate / low / high> columns per parameter
}\if{html}{\out{</div>}}
}
\description{
Summarise list of comet-fit results into a tidy data-frame
}
