% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prepare_comet_data}
\alias{prepare_comet_data}
\title{prepare_comet_data}
\usage{
prepare_comet_data(dataset, sample_name)
}
\arguments{
\item{dataset}{input dataframe storing comet kinetic data, requires columns: Sample and F_t values at different time points}

\item{sample_name}{input individual sample name}
}
\value{
a formatted dataframe with time, F_t, and sample name
}
\description{
Prepare comet assay data for one sample.
}
