% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{RE_to_zscore}
\alias{RE_to_zscore}
\title{RE_to_zscore}
\usage{
RE_to_zscore(dataset, var_inverse = NULL)
}
\arguments{
\item{dataset}{input dataframe storing raw reporter expression data - make sure the data is appropriately ordered}

\item{var_inverse}{input pathway names with inverse relationship between raw reporter expression and pathway activity}
}
\value{
list with standardized data and mean/sd
}
\description{
Log transform and z-score normalize FM-HCR data.
}
